Bug fix/split sync testcase (#21599)
* split the testcase into more handy bits for instrumented runs

* cleanup

* more buckets
